+++ test/token/LifTokenTest.js
